requirements.txt
gradio
gtts
pydub
requests
python-dotenv
elevenlabs
